refactor(user): wire the edit route in place of the name-change route
